+++ ocrd_tesserocr/recognize.py
@@ -11,6 +11,7 @@ from tesserocr import (
     Orientation,
     WritingDirection,
     TextlineOrder,
+    tesseract_version,
     PyTessBaseAPI, get_languages as get_languages_)
 
 from ocrd_utils import (
@@ -72,7 +73,7 @@ class TesserocrRecognize(Processor):
 
     def __init__(self, *args, **kwargs):
         kwargs['ocrd_tool'] = OCRD_TOOL['tools'][TOOL]
-        kwargs['version'] = OCRD_TOOL['version']
+        kwargs['version'] = OCRD_TOOL['version'] + ' (' + tesseract_version().split('\n')[0] + ')'
         super(TesserocrRecognize, self).__init__(*args, **kwargs)
         
         if hasattr(self, 'workspace'):
@@ -92,7 +93,9 @@ class TesserocrRecognize(Processor):
         
         Set up Tesseract to recognise each segment's image (either from
         AlternativeImage or cropping the bounding box rectangle and masking
-        it from the polygon outline) with the appropriate mode and ``model``.
+        it from the polygon outline) with the appropriate segmentation mode
+        and ``model``. (If no ``model`` is given, only layout analysis will
+        be performed.)
         
         Next, if there still is a gap between the current level in the PAGE hierarchy
         and the requested ``textequiv_level``, then iterate down the result hierarchy,
@@ -100,7 +103,8 @@ class TesserocrRecognize(Processor):
         text line order, reading direction and orientation at the region/table level).
         
         Then, at ``textequiv_level``, remove any existing TextEquiv, unless
-        ``overwrite_text`` is false, and add text and confidence results .
+        ``overwrite_text`` is false, and add text and confidence results, unless
+        ``model`` is empty.
         
         The special value ``textequiv_level=none`` behaves like ``glyph``,
         except that no actual text recognition will be performed, only
@@ -121,28 +125,33 @@ class TesserocrRecognize(Processor):
         In other words:
         - If ``segmentation_level=region``, then segment the page into regions
           (unless ``overwrite_segments=false``), else iterate existing regions.
-        - If ``textequiv_level=region``, then recognize text in the region,
-          annotate it, and continue with the next region. Otherwise...
+        - If ``textequiv_level=region``, then unless ``model`` is empty,
+          recognize text in the region and annotate it. Regardless, continue
+          with the next region. Otherwise...
         - If ``segmentation_level=cell`` or higher,
           then segment table regions into text regions (i.e. cells)
           (unless ``overwrite_segments=false``), else iterate existing cells.
-        - If ``textequiv_level=cell``, then recognize text in the cell,
-          annotate it, and continue with the next cell. Otherwise...
+        - If ``textequiv_level=cell``, then unless ``model`` is empty,
+          recognize text in the cell and annotate it. Regardless, continue
+          with the next cell. Otherwise...
         - If ``segmentation_level=line`` or higher,
           then segment text regions into text lines
           (unless ``overwrite_segments=false``), else iterate existing text lines.
-        - If ``textequiv_level=line``, then recognize text in the text lines,
-          annotate it, and continue with the next line. Otherwise...
+        - If ``textequiv_level=line``, then unless ``model`` is empty,
+          recognize text in the text lines and annotate it. Regardless, continue
+          with the next line. Otherwise...
         - If ``segmentation_level=word`` or higher,
           then segment text lines into words
           (unless ``overwrite_segments=false``), else iterate existing words.
-        - If ``textequiv_level=word``, then recognize text in the words,
-          annotate it, and continue with the next word. Otherwise...
+        - If ``textequiv_level=word``, then unless ``model`` is empty,
+          recognize text in the words and annotate it. Regardless, continue
+          with the next word. Otherwise...
         - If ``segmentation_level=glyph`` or higher,
           then segment words into glyphs
           (unless ``overwrite_segments=false``), else iterate existing glyphs.
-        - If ``textequiv_level=glyph``, then recognize text in the glyphs and
-          continue with the next glyph. Otherwise...
+        - If ``textequiv_level=glyph``, then unless ``model`` is empty,
+          recognize text in the glyphs and annotate it. Regardless, continue
+          with the next glyph. Otherwise...
         - (i.e. ``none``) annotate no text and be done.
         
         Note that ``cell`` is an _optional_ level that is only relevant for
@@ -156,7 +165,8 @@ class TesserocrRecognize(Processor):
         Thus, ``segmentation_level`` is the entry point level for layout analysis,
         and setting it to ``none`` makes this processor behave as recognition-only.
         Whereas ``textequiv_level`` selects the exit point level for segmentation,
-        and setting it to ``none`` makes this processor behave as segmentation-only.
+        and setting it to ``none`` makes this processor behave as segmentation-only,
+        as does omitting ``model``.
         
         All segments above ``segmentation_level`` must already exist, and
         no segments below ``textequiv_level`` will be newly created.
@@ -170,9 +180,18 @@ class TesserocrRecognize(Processor):
         (This is more precise, but due to some path representation errors does
         not always yield accurate/valid polygons.)
         
+        If ``shrink_polygons``, then during segmentation (on any level), query Tesseract
+        for all symbols/glyphs of each segment and calculate the convex hull for them.
+        Annotate the resulting polygon instead of the coarse bounding box.
+        (This is more precise and helps avoid overlaps between neighbours, especially
+        when not segmenting all levels at once.)
+        
         If ``sparse_text``, then during region segmentation, attempt to find
         single-line text blocks in no particular order (Tesseract's page segmentation
         mode ``SPARSE_TEXT``).
+        
+        If ``tesseract_parameters`` is given, setup each of its key-value pairs as
+        run-time parameters in Tesseract.
         
         Finally, produce new output files by serialising the resulting hierarchy.
         """
@@ -190,15 +209,15 @@ class TesserocrRecognize(Processor):
             for sub_model in model.split('+'):
                 if sub_model not in get_languages()[1]:
                     raise Exception("configured model " + sub_model + " is not installed")
+                self.logger.info("Using model '%s' in %s for recognition at the %s level",
+                                 model, get_languages()[0], outlevel)
         
         with PyTessBaseAPI(path=TESSDATA_PREFIX, lang=model) as tessapi:
-            self.logger.info("Using model '%s' in %s for recognition at the %s level",
-                             model, get_languages()[0], outlevel)
             if outlevel == 'glyph':
                 # populate GetChoiceIterator() with LSTM models, too:
                 tessapi.SetVariable("lstm_choice_mode", "2") # aggregate symbols
                 tessapi.SetVariable("lstm_choice_iterations", "15") # squeeze out more best paths
-            if outlevel != 'none':
+            if outlevel != 'none' and self.parameter.get('model', ''):
                 # TODO: maybe warn/raise when illegal combinations or characters not in the model unicharset?
                 if self.parameter['char_whitelist']:
                     tessapi.SetVariable("tessedit_char_whitelist", self.parameter['char_whitelist'])
@@ -251,6 +270,9 @@ class TesserocrRecognize(Processor):
             # lstm_use_matrix 1
             # user_words_file
             # user_patterns_file
+            tesseract_params = self.parameter['tesseract_parameters']
+            for variable in tesseract_params:
+                tessapi.SetVariable(variable, tesseract_params[variable])
             for (n, input_file) in enumerate(self.input_files):
                 page_id = input_file.pageId or input_file.ID
                 self.logger.info("INPUT FILE %i / %s", n, page_id)
@@ -307,7 +329,7 @@ class TesserocrRecognize(Processor):
                     page.set_ReadingOrder(None)
                     # prepare Tesseract
                     if self.parameter['find_tables']:
-                        if outlevel == 'region':
+                        if outlevel == 'region' and self.parameter.get('model', ''):
                             raise Exception("When segmentation_level is region and find_tables is enabled, textequiv_level must be at least cell, because text results cannot be annotated on tables directly.")
                         tessapi.SetVariable("textord_tabfind_find_tables", "1") # (default)
                         # this should yield additional blocks within the table blocks
@@ -323,7 +345,7 @@ class TesserocrRecognize(Processor):
                     tessapi.SetPageSegMode(PSM.SPARSE_TEXT
                                            if self.parameter['sparse_text']
                                            else PSM.AUTO)
-                    if outlevel == 'none':
+                    if outlevel == 'none' or not self.parameter.get('model', ''):
                         self.logger.debug("Detecting regions in page '%s'", page_id)
                         tessapi.AnalyseLayout()
                     else:
@@ -351,8 +373,13 @@ class TesserocrRecognize(Processor):
                     self.logger.warning("Page '%s' contains no text regions (but segmentation is off)",
                                             page_id)
                 
-                if outlevel != 'none':
+                # post-processing
+                # bottom-up text concatenation
+                if outlevel != 'none' and self.parameter.get('model', ''):
                     page_update_higher_textequiv_levels(outlevel, pcgts)
+                # bottom-up polygonal outline projection
+                # if inlevel != 'none' and self.parameter['shrink_polygons']:
+                #     page_shrink_higher_coordinate_levels(inlevel, outlevel, pcgts)
                 
                 file_id = make_file_id(input_file, self.output_file_grp)
                 pcgts.set_pcGtsId(file_id)
@@ -387,12 +414,12 @@ class TesserocrRecognize(Processor):
         # (which would also give raw coordinates),
         # except we are also interested in the iterator's BlockType() here,
         # and its BlockPolygon()
-        for it in iterate_level(result_it, RIL.BLOCK):
+        for i, it in enumerate(iterate_level(result_it, RIL.BLOCK)):
             # (padding will be passed to both BoundingBox and GetImage)
             # (actually, Tesseract honours padding only on the left and bottom,
             #  whereas right and top are increased less!)
             # TODO: output padding can create overlap between neighbours; at least find polygonal difference
-            x0y0x1y1 = it.BoundingBox(RIL.BLOCK, padding=self.parameter['padding'])
+            bbox = it.BoundingBox(RIL.BLOCK, padding=self.parameter['padding'])
             # sometimes these polygons are not planar, which causes
             # PIL.ImageDraw.Draw.polygon (and likely others as well)
             # to misbehave; however, PAGE coordinate semantics prohibit
@@ -400,8 +427,17 @@ class TesserocrRecognize(Processor):
             # (probably a bug in Tesseract itself, cf. tesseract#2826):
             if self.parameter['block_polygons']:
                 polygon = it.BlockPolygon()
+            elif self.parameter['shrink_polygons'] and not it.Empty(RIL.SYMBOL):
+                polygon = join_polygons([polygon_from_x0y0x1y1(
+                    symbol.BoundingBox(RIL.SYMBOL, padding=self.parameter['padding']))
+                                         for symbol in iterate_level(it, RIL.SYMBOL, parent=RIL.BLOCK)])
+                # simulate a RestartBlock(), not defined by Tesseract:
+                it.Begin()
+                for j, it in enumerate(iterate_level(it, RIL.BLOCK)):
+                    if i == j:
+                        break
             else:
-                polygon = polygon_from_x0y0x1y1(x0y0x1y1)
+                polygon = polygon_from_x0y0x1y1(bbox)
             xywh = xywh_from_polygon(polygon)
             polygon = coordinates_for_segment(polygon, None, page_coords)
             polygon2 = polygon_for_parent(polygon, page)
@@ -424,7 +460,7 @@ class TesserocrRecognize(Processor):
                     PT.EQUATION,
                     PT.TABLE] and (
                         xywh['w'] < 20 / 300.0*(dpi or 300) or
-                        xywh['h'] < 30 / 300.0*(dpi or 300)):
+                        xywh['h'] < 10 / 300.0*(dpi or 300)):
                 self.logger.info('Ignoring too small region: %s', points)
                 continue
             region_image_bin = it.GetBinaryImage(RIL.BLOCK)
@@ -462,13 +498,13 @@ class TesserocrRecognize(Processor):
                     region.set_type(TextTypeSimpleType.CAPTION)
                 page.add_TextRegion(region)
                 og.add_RegionRefIndexed(RegionRefIndexedType(regionRef=ID, index=index))
-                if self.parameter['textequiv_level'] == 'region':
+                if self.parameter['textequiv_level'] not in ['region', 'cell']:
+                    self._process_lines_in_region(it, region, page_coords)
+                elif self.parameter.get('model', ''):
                     region.add_TextEquiv(TextEquivType(
                         Unicode=it.GetUTF8Text(RIL.BLOCK).rstrip("\n\f"),
                         # iterator scores are arithmetic averages, too
                         conf=it.Confidence(RIL.BLOCK)/100.0))
-                else:
-                    self._process_lines_in_region(it, region, page_coords)
             elif block_type in [PT.FLOWING_IMAGE,
                                 PT.HEADING_IMAGE,
                                 PT.PULLOUT_IMAGE]:
@@ -527,7 +563,20 @@ class TesserocrRecognize(Processor):
             ril = RIL.PARA # for "cells" in PT.TABLE block
         for index, it in enumerate(iterate_level(result_it, ril)):
             bbox = it.BoundingBox(ril, padding=self.parameter['padding'])
-            polygon = polygon_from_x0y0x1y1(bbox)
+            if self.parameter['shrink_polygons'] and not it.Empty(RIL.SYMBOL):
+                polygon = join_polygons([polygon_from_x0y0x1y1(
+                    symbol.BoundingBox(RIL.SYMBOL, padding=self.parameter['padding']))
+                                         for symbol in iterate_level(it, RIL.SYMBOL, parent=ril)])
+                if ril == RIL.BLOCK:
+                    # simulate a RestartBlock(), not defined by Tesseract:
+                    it.Begin()
+                    for j, it in enumerate(iterate_level(it, RIL.BLOCK)):
+                        if index == j:
+                            break
+                else:
+                    it.RestartParagraph()
+            else:
+                polygon = polygon_from_x0y0x1y1(bbox)
             polygon = coordinates_for_segment(polygon, None, page_coords)
             polygon2 = polygon_for_parent(polygon, region)
             if polygon2 is not None:
@@ -544,14 +593,14 @@ class TesserocrRecognize(Processor):
             self._add_orientation(it, cell, page_coords)
             if rogroup:
                 rogroup.add_RegionRefIndexed(RegionRefIndexedType(regionRef=ID, index=index))
-            if self.parameter['textequiv_level'] == 'cell':
+            if self.parameter['textequiv_level'] != 'cell':
+                self._process_lines_in_region(it, cell, page_coords, parent_ril=ril)
+            elif self.parameter.get('model', ''):
                 cell.add_TextEquiv(TextEquivType(
                     Unicode=it.GetUTF8Text(ril).rstrip("\n\f"),
                     # iterator scores are arithmetic averages, too
                     conf=it.Confidence(ril)/100.0))
-            else:
-                self._process_lines_in_region(it, cell, page_coords, parent_ril=ril)
-        
+
     def _process_lines_in_region(self, result_it, region, page_coords, parent_ril=RIL.BLOCK):
         if self.parameter['sparse_text']:
             it = result_it
@@ -559,18 +608,24 @@ class TesserocrRecognize(Processor):
             line = TextLineType(id=region.id + '_line',
                                 Coords=region.get_Coords())
             region.add_TextLine(line)
-            if self.parameter['textequiv_level'] == 'line':
+            if self.parameter['textequiv_level'] != 'line':
+                self._process_words_in_line(it, line, page_coords)
+            elif self.parameter.get('model', ''):
                 # todo: consider BlankBeforeWord, SetLineSeparator
                 line.add_TextEquiv(TextEquivType(
                     Unicode=it.GetUTF8Text(RIL.TEXTLINE).rstrip("\n\f"),
                     # iterator scores are arithmetic averages, too
                     conf=it.Confidence(RIL.TEXTLINE)/100.0))
-            else:
-                self._process_words_in_line(it, line, page_coords)
             return
         for index, it in enumerate(iterate_level(result_it, RIL.TEXTLINE, parent=parent_ril)):
             bbox = it.BoundingBox(RIL.TEXTLINE, padding=self.parameter['padding'])
-            polygon = polygon_from_x0y0x1y1(bbox)
+            if self.parameter['shrink_polygons'] and not it.Empty(RIL.SYMBOL):
+                polygon = join_polygons([polygon_from_x0y0x1y1(
+                    symbol.BoundingBox(RIL.SYMBOL, padding=self.parameter['padding']))
+                                         for symbol in iterate_level(it, RIL.SYMBOL, parent=RIL.TEXTLINE)])
+                it.RestartRow()
+            else:
+                polygon = polygon_from_x0y0x1y1(bbox)
             polygon = coordinates_for_segment(polygon, None, page_coords)
             polygon2 = polygon_for_parent(polygon, region)
             if polygon2 is not None:
@@ -584,19 +639,29 @@ class TesserocrRecognize(Processor):
             self.logger.info("Detected line '%s': %s", ID, points)
             line = TextLineType(id=ID, Coords=coords)
             region.add_TextLine(line)
-            if self.parameter['textequiv_level'] == 'line':
+            if self.parameter['textequiv_level'] != 'line':
+                self._process_words_in_line(it, line, page_coords)
+            elif self.parameter.get('model', ''):
                 # todo: consider BlankBeforeWord, SetLineSeparator
                 line.add_TextEquiv(TextEquivType(
                     Unicode=it.GetUTF8Text(RIL.TEXTLINE).rstrip("\n\f"),
                     # iterator scores are arithmetic averages, too
                     conf=it.Confidence(RIL.TEXTLINE)/100.0))
-            else:
-                self._process_words_in_line(it, line, page_coords)
-    
+
     def _process_words_in_line(self, result_it, line, coords):
         for index, it in enumerate(iterate_level(result_it, RIL.WORD)):
             bbox = it.BoundingBox(RIL.WORD, padding=self.parameter['padding'])
-            polygon = polygon_from_x0y0x1y1(bbox)
+            if self.parameter['shrink_polygons'] and not it.Empty(RIL.SYMBOL):
+                polygon = join_polygons([polygon_from_x0y0x1y1(
+                    symbol.BoundingBox(RIL.SYMBOL, padding=self.parameter['padding']))
+                                         for symbol in iterate_level(it, RIL.SYMBOL, parent=RIL.WORD)])
+                # simulate a BeginWord(index), not exposed by tesserocr:
+                it.RestartRow()
+                for j, it in enumerate(iterate_level(it, RIL.WORD)):
+                    if index == j:
+                        break
+            else:
+                polygon = polygon_from_x0y0x1y1(bbox)
             polygon = coordinates_for_segment(polygon, None, coords)
             polygon2 = polygon_for_parent(polygon, line)
             if polygon2 is not None:
@@ -609,14 +674,14 @@ class TesserocrRecognize(Processor):
             self.logger.debug("Detected word '%s': %s", ID, points)
             word = WordType(id=ID, Coords=CoordsType(points=points))
             line.add_Word(word)
-            if self.parameter['textequiv_level'] == 'word':
+            if self.parameter['textequiv_level'] != 'word':
+                self._process_glyphs_in_word(it, word, coords)
+            elif self.parameter.get('model', ''):
                 word.add_TextEquiv(TextEquivType(
                     Unicode=it.GetUTF8Text(RIL.WORD),
                     # iterator scores are arithmetic averages, too
                     conf=it.Confidence(RIL.WORD)/100.0))
-            else:
-                self._process_glyphs_in_word(it, word, coords)
-    
+
     def _process_glyphs_in_word(self, result_it, word, coords):
         for index, it in enumerate(iterate_level(result_it, RIL.SYMBOL)):
             bbox = it.BoundingBox(RIL.SYMBOL, padding=self.parameter['padding'])
@@ -633,7 +698,9 @@ class TesserocrRecognize(Processor):
             #self.logger.debug("Detected glyph '%s': %s", ID, points)
             glyph = GlyphType(id=ID, Coords=CoordsType(points))
             word.add_Glyph(glyph)
-            if self.parameter['textequiv_level'] == 'glyph':
+            if self.parameter['textequiv_level'] != 'glyph':
+                pass
+            elif self.parameter.get('model', ''):
                 # glyph_text = it.GetUTF8Text(RIL.SYMBOL) # equals first choice?
                 glyph_conf = it.Confidence(RIL.SYMBOL)/100 # equals first choice?
                 #self.logger.debug('best glyph: "%s" [%f]', glyph_text, glyph_conf)
@@ -723,7 +790,7 @@ class TesserocrRecognize(Processor):
                 tessapi.SetImage(table_image)
             tessapi.SetPageSegMode(PSM.SPARSE_TEXT) # retrieve "cells"
             # TODO: we should XY-cut the sparse cells in regroup them into consistent cells
-            if self.parameter['textequiv_level'] == 'none':
+            if self.parameter['textequiv_level'] == 'none' or not self.parameter.get('model', ''):
                 self.logger.debug("Detecting cells in table '%s'", table.id)
                 tessapi.AnalyseLayout()
             else:
@@ -732,13 +799,15 @@ class TesserocrRecognize(Processor):
             self._process_cells_in_table(tessapi.GetIterator(), table, roelem, table_coords)
     
     def _process_existing_regions(self, tessapi, regions, page_image, page_coords):
+        if self.parameter['textequiv_level'] in ['region', 'cell'] and not self.parameter.get('model', ''):
+            return
         for region in regions:
             region_image, region_coords = self.workspace.image_from_segment(
                 region, page_image, page_coords)
             if not region_image.width or not region_image.height:
                 self.logger.warning("Skipping text region '%s' with zero size", region.id)
                 continue
-            if (self.parameter['textequiv_level'] != 'region' and
+            if (self.parameter['textequiv_level'] not in ['region', 'cell'] and
                 self.parameter['segmentation_level'] != 'line'):
                 pass # image not used here
             elif self.parameter['padding']:
@@ -768,7 +837,7 @@ class TesserocrRecognize(Processor):
                 if textlines:
                     self.logger.info('Removing existing text lines in region %s', region.id)
                 region.set_TextLine([])
-                if self.parameter['textequiv_level'] == 'none':
+                if self.parameter['textequiv_level'] == 'none' or not self.parameter.get('model', ''):
                     self.logger.debug("Detecting lines in region '%s'", region.id)
                     tessapi.AnalyseLayout()
                 else:
@@ -782,6 +851,8 @@ class TesserocrRecognize(Processor):
                                     region.id)
 
     def _process_existing_lines(self, tessapi, textlines, region_image, region_coords):
+        if self.parameter['textequiv_level'] == 'line' and not self.parameter.get('model', ''):
+            return
         for line in textlines:
             line_image, line_coords = self.workspace.image_from_segment(
                 line, region_image, region_coords)
@@ -820,7 +891,7 @@ class TesserocrRecognize(Processor):
                 if words:
                     self.logger.info('Removing existing words in line %s', line.id)
                 line.set_Word([])
-                if self.parameter['textequiv_level'] == 'none':
+                if self.parameter['textequiv_level'] == 'none' or not self.parameter.get('model', ''):
                     self.logger.debug("Detecting words in line '%s'", line.id)
                     tessapi.AnalyseLayout()
                 else:
@@ -837,6 +908,8 @@ class TesserocrRecognize(Processor):
                                     line.id)
 
     def _process_existing_words(self, tessapi, words, line_image, line_coords):
+        if self.parameter['textequiv_level'] == 'word' and not self.parameter.get('model', ''):
+            return
         for word in words:
             word_image, word_coords = self.workspace.image_from_segment(
                 word, line_image, line_coords)
@@ -870,7 +943,7 @@ class TesserocrRecognize(Processor):
                 if glyphs:
                     self.logger.info('Removing existing glyphs in word %s', word.id)
                 word.set_Glyph([])
-                if self.parameter['textequiv_level'] == 'none':
+                if self.parameter['textequiv_level'] == 'none' or not self.parameter.get('model', ''):
                     self.logger.debug("Detecting glyphs in word '%s'", word.id)
                     tessapi.AnalyseLayout()
                 else:
@@ -887,6 +960,8 @@ class TesserocrRecognize(Processor):
                                     word.id)
 
     def _process_existing_glyphs(self, tessapi, glyphs, word_image, word_xywh):
+        if not self.parameter.get('model', ''):
+            return
         for glyph in glyphs:
             glyph_image, _ = self.workspace.image_from_segment(
                 glyph, word_image, word_xywh)
@@ -1009,7 +1084,7 @@ def page_get_reading_order(ro, rogroup):
 def page_update_higher_textequiv_levels(level, pcgts):
     """Update the TextEquivs of all PAGE-XML hierarchy levels above ``level`` for consistency.
     
-    Starting with the hierarchy level chosen for processing,
+    Starting with the lowest hierarchy level chosen for processing,
     join all first TextEquiv.Unicode (by the rules governing the respective level)
     into TextEquiv.Unicode of the next higher level, replacing them.
     
@@ -1116,6 +1191,58 @@ def page_update_higher_textequiv_levels(level, pcgts):
             region.set_TextEquiv( # replace old, if any
                 [TextEquivType(Unicode=region_unicode, conf=region_conf)])
 
+def page_shrink_higher_coordinate_levels(maxlevel, minlevel, pcgts):
+    """Project the coordinate hull of all PAGE-XML hierarchy levels above ``minlevel`` up to ``maxlevel``.
+    
+    Starting with the lowest hierarchy level chosen for processing,
+    join all segments into a convex hull for the next higher level,
+    replacing the parent coordinates, respectively.
+    
+    Follow regions recursively, but make sure to traverse them in a depth-first strategy.
+    """
+    LOG = getLogger('processor.TesserocrRecognize')
+    page = pcgts.get_Page()
+    regions = page.get_AllRegions(classes=['Text'])
+    if minlevel != 'region':
+        for region in regions:
+            lines = region.get_TextLine()
+            if minlevel != 'line':
+                for line in lines:
+                    words = line.get_Word()
+                    if minlevel != 'word':
+                        for word in words:
+                            glyphs = word.get_Glyph()
+                            if maxlevel in ['region', 'line', 'word', 'glyph'] and glyphs:
+                                joint_polygon = join_segments(glyphs)
+                                LOG.debug("setting hull for word '%s' from %d vertices",
+                                          word.id, len(joint_polygon))
+                                word.get_Coords().set_points(points_from_polygon(joint_polygon))
+                    if maxlevel in ['region', 'line', 'word'] and words:
+                        joint_polygon = join_segments(words)
+                        LOG.debug("setting hull for line '%s' from %d vertices",
+                                  line.id, len(joint_polygon))
+                        line.get_Coords().set_points(points_from_polygon(joint_polygon))
+            if maxlevel in ['region', 'line'] and lines:
+                joint_polygon = join_segments(lines)
+                LOG.debug("setting hull for region '%s' from %d vertices",
+                          region.id, len(joint_polygon))
+                region.get_Coords().set_points(points_from_polygon(joint_polygon))
+
+def join_segments(segments):
+    return join_polygons([polygon_from_points(segment.get_Coords().points)
+                          for segment in segments])
+
+def join_polygons(polygons, extend=2):
+    # FIXME: construct concave hull / alpha shape
+    jointp = unary_union([make_valid(Polygon(polygon)).buffer(extend)
+                          for polygon in polygons]).convex_hull
+    if jointp.minimum_clearance < 1.0:
+        # follow-up calculations will necessarily be integer;
+        # so anticipate rounding here and then ensure validity
+        jointp = asPolygon(np.round(jointp.exterior.coords))
+        jointp = make_valid(jointp)
+    return jointp.exterior.coords[:-1]
+
 def pad_image(image, padding):
     # TODO: input padding can create extra edges if not binarized; at least try to smooth
     stat = ImageStat.Stat(image)
@@ -1189,13 +1316,41 @@ def make_valid(polygon):
     return polygon
 
 def iterate_level(it, ril, parent=None):
+    LOG = getLogger('processor.TesserocrRecognize')
     # improves over tesserocr.iterate_level by
     # honouring multi-level semantics so iterators
     # can be combined across levels
     if parent is None:
         parent = ril - 1
+    pos = 0
     while it and not it.Empty(ril):
         yield it
-        if ril > 0 and it.IsAtFinalElement(parent, ril):
+        # With upstream Tesseract, these assertions may fail:
+        # if ril > 0 and it.IsAtFinalElement(parent, ril):
+        #     for level in range(parent, ril):
+        #         assert it.IsAtFinalElement(parent, level), \
+        #             "level %d iterator at %d is final w.r.t. %d but level %d is not" % (
+        #                 ril, pos, parent, level)
+        # Hence the following workaround avails itself:
+        if ril > 0 and all(it.IsAtFinalElement(parent, level)
+                           for level in range(parent, ril + 1)):
             break
-        it.Next(ril)
+        if not it.Next(ril):
+            break
+        while it.Empty(ril) and not it.Empty(0):
+            # This happens when
+            # - on RIL.PARA, RIL.TEXTLINE and RIL.WORD,
+            #   empty non-text (pseudo-) blocks intervene
+            # - on RIL.SYMBOL, a word has no cblobs at all
+            #   (because they have all been rejected)
+            # We must _not_ yield these (as they have strange
+            # properties and bboxes). But most importantly,
+            # they will have met IsAtFinalElement prematurely
+            # (hence the similar loop above).
+            # Since this may happen multiple consecutive times,
+            # enclose this in a while loop.
+            LOG.warning("level %d iterator at %d needs to skip empty segment",
+                        ril, pos)
+            if not it.Next(ril):
+                break
+        pos += 1