+++ ocrd_tesserocr/segment.py
@@ -21,6 +21,7 @@ class TesserocrSegment(Processor):
             recognize_kwargs.pop('show_help', None)
             recognize_kwargs.pop('show_version', None)
             recognize_kwargs['parameter'] = self.parameter
+            recognize_kwargs['parameter']['overwrite_segments'] = True
             recognize_kwargs['parameter']['segmentation_level'] = "region"
             recognize_kwargs['parameter']['textequiv_level'] = "none"
             self.recognizer = TesserocrRecognize(self.workspace, **recognize_kwargs)
@@ -44,6 +45,12 @@ class TesserocrSegment(Processor):
         instead of bounding boxes for each region.
         (This is more precise, but due to some path representation errors does
         not always yield accurate/valid polygons.)
+        
+        If ``shrink_polygons``, then query Tesseract for all symbols/glyphs
+        of each segment and calculate the convex hull for them.
+        Annotate the resulting polygon instead of the coarse bounding box.
+        (This is more precise and helps avoid overlaps between neighbours, especially
+        when not segmenting all levels at once.)
         
         If ``sparse_text``, then attempt to find single-line text blocks only,
         in no particular order.